Implemented (===)/7 to remove useless choice points in factorial/2
(===)/7 is the equivalent of the reified (=)/3 for two couples of variables. The penultimate argument is used to know which couple was responsible for the true/false choice, so that one can use the other couple in the `then` part of the if_/3 call.

The first couple is the one used when both couples are free variables, so as to not get duplicates.

This predicate is useful for situations where we have:

if_(A = B,
    C = D,
    [else…]
)

which can be replaced by:

if_(===(A, B, C, D, T1),
    if_(T1 = true,
        C = D,
        A = B
    ),
    [else…]
)

which will at most only remove unnecessary choice points when C and D are ground compared to the (=)/3 version.